requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.1,<4.0
celery>=5.4,<6.0
redis>=5.0,<6.0
python-multipart>=0.0.9,<1.0
pydantic-settings>=2.4,<3.0
faiss-cpu>=1.8,<2.0
sentence-transformers>=3.0,<6.0
openai>=1.30,<2.0
pypdf>=4.3,<6.0
python-docx>=1.1,<2.0
numpy>=1.26,<3.0
tenacity>=8.2,<10.0
httpx>=0.27,<1.0
